+++ src/Creational/Factories/Method/Product/Product.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Method\Product;
+
+interface Product
+{
+    public function type(): string;
+}